LightNovel CRUD update

+++ src/main/java/com/my_project/LightNovel_web_backend/repository/GenreRepository.java
@@ -3,5 +3,5 @@ package com.my_project.LightNovel_web_backend.repository;
 import com.my_project.LightNovel_web_backend.entity.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
 }
